Enhance rate limiter algorithms with logging and error handling; update memory.py to use int for tokens; add edge-case tests for validation and failure policies.

+++ src/algorithms/memory.py
@@ -2,7 +2,7 @@ from typing import TypedDict
 
 
 class TokenBucketState(TypedDict):
-    tokens: float
+    tokens: int
     last_refill_timestamp: float
